Move DataAccess from access.py to dataaccess.py.
